requirements.txt
# Data handling
numpy
pandas
torch

# LLM
chatarena
openai
cohere==4.53
# TODO add other LLM libraires here

# Visualisations
matplotlib
seaborn

# OS stuff
python-dotenv==1.0.1
unidecode

# any other librairies...